+++ mycodo/mycodo_flask/forms/forms_input.py
@@ -18,7 +18,7 @@ from wtforms import widgets
 from wtforms.validators import DataRequired
 from wtforms.widgets.html5 import NumberInput
 
-from mycodo.config_translations import TOOLTIPS_INPUT
+from mycodo.config_translations import TOOLTIPS_SETTINGS
 from mycodo.mycodo_flask.utils.utils_general import generate_form_input_list
 from mycodo.utils.inputs import parse_input_information
 
@@ -69,7 +69,7 @@ class InputMod(FlaskForm):
         validators=[DataRequired()]
     )
     period = DecimalField(
-        TOOLTIPS_INPUT['period']['title'],
+        TOOLTIPS_SETTINGS['period']['title'],
         validators=[DataRequired(),
                     validators.NumberRange(
                         min=5.0,
@@ -81,7 +81,8 @@ class InputMod(FlaskForm):
     ftdi_location = StringField(lazy_gettext('FTDI Device'))  # FTDI device location type
     uart_location = StringField(lazy_gettext('UART Device'))  # UART device location type
     i2c_location = StringField(lazy_gettext('I<sup>2</sup>C Address'))  # I2C device location type
-    gpio_location = IntegerField(lazy_gettext('GPIO Pin'))  # GPIO device location type
+    gpio_location = IntegerField(
+        TOOLTIPS_SETTINGS['gpio_location']['title'])
 
     i2c_bus = IntegerField(
         lazy_gettext('I<sup>2</sup>C Bus'), widget=NumberInput())
@@ -99,47 +100,47 @@ class InputMod(FlaskForm):
         lazy_gettext('Sensitivity'), widget=NumberInput())
 
     measurements_enabled = SelectMultipleField(
-        TOOLTIPS_INPUT['measurements_enabled']['title'])
+        TOOLTIPS_SETTINGS['measurements_enabled']['title'])
 
     # Server options
     host = StringField(lazy_gettext('Host'))
     port = IntegerField(
-        TOOLTIPS_INPUT['port']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['port']['title'], widget=NumberInput())
     times_check = IntegerField(
-        TOOLTIPS_INPUT['times_check']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['times_check']['title'], widget=NumberInput())
     deadline = IntegerField(
-        TOOLTIPS_INPUT['deadline']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['deadline']['title'], widget=NumberInput())
 
     # Linux Command
     cmd_command = StringField(
-        TOOLTIPS_INPUT['cmd_command']['title'])
+        TOOLTIPS_SETTINGS['cmd_command']['title'])
 
     # MAX chip options
     thermocouple_type = StringField(
-        TOOLTIPS_INPUT['thermocouple_type']['title'])
+        TOOLTIPS_SETTINGS['thermocouple_type']['title'])
     ref_ohm = IntegerField(
-        TOOLTIPS_INPUT['ref_ohm']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['ref_ohm']['title'], widget=NumberInput())
 
     # SPI Communication
     pin_clock = IntegerField(
-        TOOLTIPS_INPUT['pin_clock']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['pin_clock']['title'], widget=NumberInput())
     pin_cs = IntegerField(
-        TOOLTIPS_INPUT['pin_cs']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['pin_cs']['title'], widget=NumberInput())
     pin_mosi = IntegerField(
-        TOOLTIPS_INPUT['pin_mosi']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['pin_mosi']['title'], widget=NumberInput())
     pin_miso = IntegerField(
-        TOOLTIPS_INPUT['pin_miso']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['pin_miso']['title'], widget=NumberInput())
 
     # Bluetooth Communication
     bt_adapter = StringField(lazy_gettext('BT Adapter (hci[X])'))
 
     # ADC
     adc_gain = IntegerField(
-        TOOLTIPS_INPUT['adc_gain']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['adc_gain']['title'], widget=NumberInput())
     adc_resolution = IntegerField(
-        TOOLTIPS_INPUT['adc_resolution']['title'], widget=NumberInput())
+        TOOLTIPS_SETTINGS['adc_resolution']['title'], widget=NumberInput())
     adc_sample_speed = StringField(
-        TOOLTIPS_INPUT['adc_sample_speed']['title'])
+        TOOLTIPS_SETTINGS['adc_sample_speed']['title'])
 
     switch_edge = StringField(lazy_gettext('Edge'))
     switch_bounce_time = IntegerField(
@@ -148,30 +149,30 @@ class InputMod(FlaskForm):
         lazy_gettext('Reset Period'), widget=NumberInput())
 
     # Pre-Output
-    pre_output_id = StringField(TOOLTIPS_INPUT['pre_output_id']['title'])
+    pre_output_id = StringField(TOOLTIPS_SETTINGS['pre_output_id']['title'])
     pre_output_duration = DecimalField(
-        TOOLTIPS_INPUT['pre_output_duration']['title'],
+        TOOLTIPS_SETTINGS['pre_output_duration']['title'],
         validators=[validators.NumberRange(
             min=0,
             max=86400
         )],
         widget=NumberInput(step='any')
     )
-    pre_output_during_measure = BooleanField(TOOLTIPS_INPUT['pre_output_during_measure']['title'])
+    pre_output_during_measure = BooleanField(TOOLTIPS_SETTINGS['pre_output_during_measure']['title'])
 
     # RPM/Signal
     weighting = DecimalField(
-        TOOLTIPS_INPUT['weighting']['title'],
+        TOOLTIPS_SETTINGS['weighting']['title'],
         widget=NumberInput(step='any'))
     rpm_pulses_per_rev = DecimalField(
-        TOOLTIPS_INPUT['rpm_pulses_per_rev']['title'],
+        TOOLTIPS_SETTINGS['rpm_pulses_per_rev']['title'],
         widget=NumberInput(step='any'))
     sample_time = DecimalField(
-        TOOLTIPS_INPUT['sample_time']['title'],
+        TOOLTIPS_SETTINGS['sample_time']['title'],
         widget=NumberInput(step='any'))
 
     # SHT options
-    sht_voltage = StringField(TOOLTIPS_INPUT['sht_voltage']['title'])
+    sht_voltage = StringField(TOOLTIPS_SETTINGS['sht_voltage']['title'])
 
     input_mod = SubmitField(lazy_gettext('Save'))
     input_delete = SubmitField(lazy_gettext('Delete'))
@@ -190,26 +191,26 @@ class InputMeasurementMod(FlaskForm):
         validators=[DataRequired()]
     )
     select_measurement_unit = StringField(
-        TOOLTIPS_INPUT['select_measurement_unit']['title'])
+        TOOLTIPS_SETTINGS['select_measurement_unit']['title'])
 
     scale_from_min = DecimalField(
-        TOOLTIPS_INPUT['scale_from_min']['title'],
+        TOOLTIPS_SETTINGS['scale_from_min']['title'],
         widget=NumberInput(step='any'))
     scale_from_max = DecimalField(
-        TOOLTIPS_INPUT['scale_from_max']['title'],
+        TOOLTIPS_SETTINGS['scale_from_max']['title'],
         widget=NumberInput(step='any'))
     scale_to_min = DecimalField(
-        TOOLTIPS_INPUT['scale_to_min']['title'],
+        TOOLTIPS_SETTINGS['scale_to_min']['title'],
         widget=NumberInput(step='any'))
     scale_to_max = DecimalField(
-        TOOLTIPS_INPUT['scale_to_max']['title'],
+        TOOLTIPS_SETTINGS['scale_to_max']['title'],
         widget=NumberInput(step='any'))
     invert_scale = BooleanField(
-        TOOLTIPS_INPUT['invert_scale']['title'])
+        TOOLTIPS_SETTINGS['invert_scale']['title'])
 
     rescaled_measurement_unit = StringField(
         lazy_gettext('Rescaled Measurement'))
     convert_to_measurement_unit = StringField(
-        TOOLTIPS_INPUT['convert_to_measurement_unit']['title'])
+        TOOLTIPS_SETTINGS['convert_to_measurement_unit']['title'])
 
     input_measurement_mod = SubmitField(lazy_gettext('Save'))