+++ mycodo/mycodo_flask/routes_page.py
@@ -49,7 +49,7 @@ from mycodo.config import RESTORE_LOG_FILE
 from mycodo.config import UPGRADE_LOG_FILE
 from mycodo.config import USAGE_REPORTS_PATH
 from mycodo.config_devices_units import MEASUREMENTS
-from mycodo.config_translations import TOOLTIPS_INPUT
+from mycodo.config_translations import TOOLTIPS_SETTINGS
 from mycodo.databases.models import Actions
 from mycodo.databases.models import AlembicVersion
 from mycodo.databases.models import Camera
@@ -1544,7 +1544,7 @@ def page_output():
                            output_info=OUTPUT_INFO,
                            output=output,
                            output_templates=output_templates,
-                           tooltips_input=TOOLTIPS_INPUT,
+                           tooltips_settings=TOOLTIPS_SETTINGS,
                            user=user)
 
 
@@ -1787,7 +1787,7 @@ def page_data():
                            table_device_measurements=DeviceMeasurements,
                            table_input=Input,
                            table_math=Math,
-                           tooltips_input=TOOLTIPS_INPUT,
+                           tooltips_settings=TOOLTIPS_SETTINGS,
                            user=user,
                            devices_1wire_ow_shell=devices_1wire_ow_shell,
                            devices_1wire_w1thermsensor=devices_1wire_w1thermsensor)